+++ packages/core/src/themes/defaultTheme.ts
@@ -60,7 +60,7 @@ export const defaultTheme: CompleteThemeSpec = {
         default: {
             stroke: '#222222',
             strokeWidth: 2,
-            fill: 'transparent',
+            fillOpacity: 0,
         },
     },
     polygon: {